requirements.txt
PyGithub>=2.3.0
PyYAML>=6.0.1
langchain>=0.3.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
langchain-huggingface>=0.1.0
langchain-openai>=0.2.0
langchain-groq>=0.1.0
langchain-anthropic>=0.2.0
langchain-chroma>=0.1.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langgraph>=0.2.0
chromadb>=0.5.0
faiss-cpu>=1.8.0
pydantic>=2.0.0
